requirements.txt
numpy==1.26.4
pandas==2.2.2
protobuf==5.27.0
python-dotenv==1.0.1
streamlit==1.29.0
scikit-learn=1.2.1
